app/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
boto3==1.34.57
